decouple genericapiserver from non-generic authenticator
