--- pkg/apiserver/authenticator/OWNERS
@@ -1,10 +1,0 @@
-reviewers:
-- smarterclayton
-- deads2k
-- liggitt
-- dims
-- spxtr
-- errordeveloper
-- resouer
-- cjcullen
-- ericchiang
