+++ pkg/genericapiserver/BUILD
@@ -34,7 +34,6 @@ go_library(
         "//pkg/apimachinery/registered:go_default_library",
         "//pkg/apis/meta/v1:go_default_library",
         "//pkg/apiserver:go_default_library",
-        "//pkg/apiserver/authenticator:go_default_library",
         "//pkg/apiserver/filters:go_default_library",
         "//pkg/apiserver/openapi:go_default_library",
         "//pkg/apiserver/request:go_default_library",
@@ -44,6 +43,7 @@ go_library(
         "//pkg/auth/handlers:go_default_library",
         "//pkg/auth/user:go_default_library",
         "//pkg/client/restclient:go_default_library",
+        "//pkg/genericapiserver/authenticator:go_default_library",
         "//pkg/genericapiserver/authorizer:go_default_library",
         "//pkg/genericapiserver/filters:go_default_library",
         "//pkg/genericapiserver/mux:go_default_library",
