+++ test/integration/framework/BUILD
@@ -30,7 +30,6 @@ go_library(
         "//pkg/apis/policy/v1alpha1:go_default_library",
         "//pkg/apis/rbac/v1alpha1:go_default_library",
         "//pkg/apis/storage/v1beta1:go_default_library",
-        "//pkg/apiserver/authenticator:go_default_library",
         "//pkg/auth/authenticator:go_default_library",
         "//pkg/auth/authorizer:go_default_library",
         "//pkg/auth/authorizer/union:go_default_library",
@@ -44,6 +43,7 @@ go_library(
         "//pkg/controller/replication:go_default_library",
         "//pkg/generated/openapi:go_default_library",
         "//pkg/genericapiserver:go_default_library",
+        "//pkg/genericapiserver/authenticator:go_default_library",
         "//pkg/genericapiserver/authorizer:go_default_library",
         "//pkg/kubectl:go_default_library",
         "//pkg/kubelet/client:go_default_library",
